Solve 2023-10
For part 1, the distance to the farthest point is the length of the loop
halved.
For part 2, I used that a point is "inside" a loop if any line from the
point to the outside crosses the loop an odd number of times.

My first time on the global leaderboard!